[main] translate solve function to python
First, I studied the algorithm and tried to translate the code line by
line. After getting quite far with that, there were still some gnarly
situations. I gave myself limited time to bugfix and then resort to
ChatGPT for a limited time. The initial ChatGPT translation was worse
than I got so far, but after feeding each of the errors, it realised all
the problems one by one and solved them.

In conclusion, we have to make sure we always convert to a string before
`eval` because of differences between Python and JavaScript. Also, we
have to care about division by 0, which throws an Exception.